Initial rollout of Hartley method of Hilbert transform. Add initial GPL statement.
The way the class calculates coefficients may be refactored, but it’s
about as efficient as it can be right now. The less efficient, but more
explicit versions of methods are still there (original *ar method for
Hilbert which does both re and im together, and the convolution method
of HilbertHRe preserved as *arConv).